+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# BetterBags - BoE & BoA Changelog
+# BetterBags - BoE, WoE & BoA Changelog
 All notable changes to this project will be documented in this file. Be aware that the [Unreleased] features are not yet available in the official tagged builds.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).